+++ src/isa/mips32/intr.c
@@ -4,7 +4,7 @@
 
 #define EX_ENTRY 0x80000180
 
-void raise_intr(uint8_t NO, vaddr_t epc) {
+void raise_intr(uint32_t NO, vaddr_t epc) {
   /* TODO: Trigger an interrupt/exception with ``NO''.
    * That is, use ``NO'' to index the IDT.
    */
@@ -30,6 +30,6 @@ bool isa_query_intr(void) {
 
 jmp_buf intr_buf;
 
-void longjmp_raise_intr(uint8_t NO) {
+void longjmp_raise_intr(uint32_t NO) {
   longjmp(intr_buf, NO + 1);
 }
