+++ src/isa/riscv32/intr.c
@@ -1,6 +1,6 @@
 #include "rtl/rtl.h"
 
-void raise_intr(uint8_t NO, vaddr_t epc) {
+void raise_intr(uint32_t NO, vaddr_t epc) {
   /* TODO: Trigger an interrupt/exception with ``NO''.
    * That is, use ``NO'' to index the IDT.
    */
