+++ src/isa/x86/exec/system.c
@@ -26,7 +26,7 @@ make_EHelper(mov_cr2r) {
 }
 
 make_EHelper(int) {
-  void raise_intr(uint8_t, vaddr_t);
+  void raise_intr(uint32_t, vaddr_t);
   raise_intr(id_dest->val, *pc);
 
   print_asm("int %s", id_dest->str);
